Add validation in Service and fixed delete bug

+++ NamespaceCBlurred-Backend/NamespaceCBlurred.Business/Services/SoundService.cs
@@ -1,4 +1,6 @@
-﻿using AutoMapper;
+﻿using System.ComponentModel.DataAnnotations;
+using AutoMapper;
+using Microsoft.IdentityModel.Tokens;
 using NamespaceCBlurred.Business.Models;
 using NamespaceCBlurred.Business.Services.Interfaces;
 using NamespaceCBlurred.Data.Models;
@@ -17,8 +19,23 @@ namespace NamespaceCBlurred.Business.Services
             this.mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
         }
 
+        private static bool ValidSoundModel(SoundForAddUpdateModel soundModel)
+        {
+            if (soundModel.Name.IsNullOrEmpty() || soundModel.SoundFilePath.IsNullOrEmpty())
+            {
+                return false;
+            }
+
+            return true;
+        }
+
         public async Task<Sound> AddSound(SoundForAddUpdateModel soundModel)
         {
+            if (!ValidSoundModel(soundModel))
+            {
+                throw new ValidationException("Invalid sound data.");
+            }
+
             var sound = mapper.Map<Sound>(soundModel);
 
             int id = await soundRepository.AddSound(sound);
@@ -29,6 +46,11 @@ namespace NamespaceCBlurred.Business.Services
 
         public async Task<bool> DeleteSound(int soundId)
         {
+            if (soundId < 0)
+            {
+                throw new ValidationException("Invalid sound id.");
+            }
+
             return await soundRepository.DeleteSound(soundId);
         }
 
@@ -39,11 +61,26 @@ namespace NamespaceCBlurred.Business.Services
 
         public async Task<Sound?> GetSoundById(int soundId)
         {
+            if (soundId < 0)
+            {
+                throw new ValidationException("Invalid sound id.");
+            }
+
             return await soundRepository.GetSoundById(soundId);
         }
 
         public async Task<bool> UpdateSound(int soundId, SoundForAddUpdateModel soundModel)
         {
+            if (soundId < 0)
+            {
+                throw new ValidationException("Invalid sound id.");
+            }
+
+            if (!ValidSoundModel(soundModel))
+            {
+                throw new ValidationException("Invalid sound data.");
+            }
+
             return await soundRepository.UpdateSound(soundId, mapper.Map<Sound>(soundModel));
         }
     }
